Remove unnecessary level check.
The check is reached only if "resolve_nameref" was called with a
non-NULL "stop" from "createparam" or "setscope" and if "pm" isn't a
reference.

The check is never reached for calls from "createparam" because these
are only made for references to not-yet-defined variables.

For "setscope", only references matter as only these can create self
references. Therefore, it makes no difference whether "pm" or NULL is
returned.