Fix rebound refs to behave the same as newly bound ones.
This change also has for effect to detect and signal incidental
reference loops (i.e., reference loops created when a reference is
rebound to a new variable on scope exit) as soon as the loop is
created rather than only when it's first accessed.

The new rebinding strategy implemented here was discussed as the all
the way up strategy.